이민하 : id 타입 통일, 서버내 사용자 목록 보이기, 멤버버튼

+++ Models/Dtos/AuthDto.cs
@@ -40,7 +40,7 @@ namespace isegoria_wpf.Models.Dtos
         );
 
         public record UserInfo(
-            [property: JsonPropertyName("id")] int Id,
+            [property: JsonPropertyName("id")] long Id,
             [property: JsonPropertyName("username")] string Username,
             [property: JsonPropertyName("email")] string Email,
             [property: JsonPropertyName("avatarUrl")] string? AvatarUrl,
